+++ sdk/storage/azure-storage-blobs/test/block_blob_client_test.cpp
@@ -7,7 +7,7 @@
 #include <random>
 #include <vector>
 
-#include <azure/core/md5.hpp>
+#include <azure/core/cryptography/hash.hpp>
 #include <azure/storage/common/crypt.hpp>
 #include <azure/storage/common/file_io.hpp>
 
@@ -121,11 +121,12 @@ namespace Azure { namespace Storage { namespace Test {
     auto res = m_blockBlobClient->Download(options);
     ASSERT_TRUE(res->TransactionalContentHash.HasValue());
     EXPECT_EQ(res->TransactionalContentHash.GetValue().Algorithm, HashAlgorithm::Md5);
-    EXPECT_EQ(
-        res->TransactionalContentHash.GetValue().Value,
-        Azure::Core::Md5::Hash(
-            std::vector<uint8_t>(m_blobContent.data(), m_blobContent.data() + downloadLength)));
-
+    {
+      Azure::Core::Cryptography::Md5Hash instance;
+      EXPECT_EQ(
+          res->TransactionalContentHash.GetValue().Value,
+          instance.Final(m_blobContent.data(), downloadLength));
+    }
     options.RangeHashAlgorithm = HashAlgorithm::Crc64;
     res = m_blockBlobClient->Download(options);
     ASSERT_TRUE(res->TransactionalContentHash.HasValue());