+++ sdk/storage/azure-storage-files-shares/test/share_file_client_test.cpp
@@ -7,7 +7,7 @@
 #include <chrono>
 #include <future>
 
-#include <azure/core/md5.hpp>
+#include <azure/core/cryptography/hash.hpp>
 #include <azure/storage/common/crypt.hpp>
 #include <azure/storage/common/file_io.hpp>
 #include <azure/storage/common/storage_common.hpp>
@@ -603,9 +603,8 @@ namespace Azure { namespace Storage { namespace Test {
 
   static std::vector<uint8_t> Hash(const std::string& data)
   {
-    const uint8_t* ptr = reinterpret_cast<const uint8_t*>(data.data());
-    std::vector<uint8_t> v(ptr, ptr + data.length());
-    return Azure::Core::Md5::Hash(v);
+    Azure::Core::Cryptography::Md5Hash instance;
+    return instance.Final(reinterpret_cast<const uint8_t*>(data.data()), data.size());
   }
 
   TEST_F(FileShareFileClientTest, RangeUploadDownload)
@@ -648,7 +647,8 @@ namespace Azure { namespace Storage { namespace Test {
     {
       // MD5 works.
       memBodyStream.Rewind();
-      auto md5 = Azure::Core::Md5::Hash(rangeContent);
+      Azure::Core::Cryptography::Md5Hash instance;
+      auto md5 = instance.Final(rangeContent.data(), rangeContent.size());
       auto invalidMd5 = Hash(std::string("This is garbage."));
       auto fileClient
           = m_shareClient->GetRootDirectoryClient().GetFileClient(LowercaseRandomString(10));